refactor(workflow): Rename `core.workflow.repository` to `core.workflow.repositories`
