+++ api/tests/unit_tests/repositories/workflow_node_execution/test_sqlalchemy_repository.py
@@ -16,7 +16,7 @@ from core.repositories import SQLAlchemyWorkflowNodeExecutionRepository
 from core.workflow.entities.node_entities import NodeRunMetadataKey
 from core.workflow.entities.workflow_node_execution import NodeExecution, NodeExecutionStatus
 from core.workflow.nodes.enums import NodeType
-from core.workflow.repository.workflow_node_execution_repository import OrderConfig
+from core.workflow.repositories.workflow_node_execution_repository import OrderConfig
 from models.account import Account, Tenant
 from models.workflow import WorkflowNodeExecution, WorkflowNodeExecutionStatus, WorkflowNodeExecutionTriggeredFrom
